requirements.txt
requests
Pillow
customtkinter
pyinstaller
pywin32
